+++ src/Components/Landing-Page/Navbar.js
@@ -26,7 +26,7 @@ const Navbar = () => {
         <div></div>
     
       <div className="navbar-right">
-        <Link to="/">Home</Link>
+        <Link to="/landing">Home</Link>
         <Link to="/signup">Signup</Link>
         <Link to="/about">About Us</Link>
         <Link to="/contact">Contact Us</Link>
